requirements.txt
amqp==5.3.1
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.15.1
channels==4.2.0
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.8.0rc2
Django==4.2.16
django-cors-headers==3.10.1
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
idna==3.8
importlib-metadata==1.7.0
importlib-resources==5.4.0
itypes==1.2.0
Jinja2==3.0.3
jsonfield==3.1.0
kombu==5.4.2
MarkupSafe==2.0.1
oauthlib==3.2.2
prompt_toolkit==3.0.48
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
requests==2.27.1
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
swapper==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
zipp==3.6.0
